requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
openai==1.55.3
httpx==0.27.2
beautifulsoup4==4.12.3
requests==2.31.0
pydantic==2.6.1
python-multipart==0.0.9
duckduckgo-search==4.1.1
markdown==3.5.2 